requirements.txt
# Existing dependencies
requests>=2.31.0

# MongoDB Atlas integration
pymongo[srv]>=4.6.0
dnspython>=2.4.2

# Email service
sendgrid>=6.11.0

# Environment variables
python-dotenv>=1.0.0

# Security
cryptography>=41.0.0
pyjwt>=2.8.0
bcrypt>=4.1.2
